Create find_dividend_index
Question:

Your task is to find and return an integer value representing the index of the dividend if present in the array. If the dividend is not found, return -1.

Problem Statement:
Given a mathematical rule that defines the dividend as:

Dividend=Divisor×Quotient+Remainder
You need to check if there exists a dividend in the array that satisfies this equation.

Input Specification:
input1: An integer array A, containing possible dividends.
input2: An integer D, representing the divisor.
input3: An integer Q, representing the quotient.
input4: An integer R, representing the remainder.
input5: An integer N, representing the length of the array A.
Output Specification:
Return an integer value representing the index of the dividend if it is present in the array.
If the dividend is not found, return -1.